refactor: 拆分guide界面mini灵动到小组件到单独的模块
